index.xhtml show data off database

+++ DeveloperGomaneh/src/java/Images.java
@@ -60,11 +60,14 @@ public class Images extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         String filename = request.getPathInfo().substring(1);
+
         File file = new File("/opt/uploads", filename);
-        response.setHeader("Content-Type", getServletContext().getMimeType(filename));
-        response.setHeader("Content-Length", String.valueOf(file.length()));
-        response.setHeader("Content-Disposition", "inline; filename=\"" + filename + "\"");
-        Files.copy(file.toPath(), response.getOutputStream());
+        if (file.exists()) {
+            response.setHeader("Content-Type", getServletContext().getMimeType(filename));
+            response.setHeader("Content-Length", String.valueOf(file.length()));
+            response.setHeader("Content-Disposition", "inline; filename=\"" + filename + "\"");
+            Files.copy(file.toPath(), response.getOutputStream());
+        }
     }
 
     /**
